Safari-609.1.11.1/LayoutTests/svg/dynamic-updates/SVGSVGElement-dom-requiredFeatures.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGSVGElement-dom-requiredFeatures.js
// [Expected rendering result] a series of PASS messages

createSVGTestCase();

var svgElement = rootSVGElement;

function repaintTest() {
    debug("Check that SVGSVGElement is initially displayed");
    shouldBeEqualToString("document.defaultView.getComputedStyle(svgElement, null).display", "inline");
    debug("Check that setting requiredFeatures to something invalid makes it not render");
    svgElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#BogusFeature");
    shouldBeEqualToString("document.defaultView.getComputedStyle(svgElement, null).display", "");
    debug("Check that setting requiredFeatures to something valid makes it render again");
    svgElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#Shape");
    shouldBeEqualToString("document.defaultView.getComputedStyle(svgElement, null).display", "inline");
    debug("Check that adding something valid to requiredFeatures keeps rendering the element");
    svgElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#Gradient");
    shouldBeEqualToString("document.defaultView.getComputedStyle(svgElement, null).display", "inline");
    debug("Check that adding something invalid to requiredFeatures makes it not render");
    svgElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#BogusFeature");
    shouldBeEqualToString("document.defaultView.getComputedStyle(svgElement, null).display", "");

    completeTest();
}


var successfullyParsed = true;
</script>
</body>
</html>
